Correção de alguns dos exercícios

+++ desafios py/desafio-090.py
@@ -1,6 +1,6 @@
 aluno = dict()
 aluno['Nome'] = str(input('Digite o nome do aluno: '))
-aluno['Média'] = int(input(f'Digite a média de {aluno["Nome"]}: '))
+aluno['Média'] = float(input(f'Digite a média de {aluno["Nome"]}: '))
 if aluno['Média'] >= 7:
     aluno['Situação'] = 'Aprovado'
 else:
